requirements.txt
# Core dependencies
openai>=1.0.0
google-generativeai>=0.3.0
structlog>=23.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Template rendering and config
Jinja2>=3.1.0
PyYAML>=6.0.0

# Web framework (for dashboard)
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
SQLAlchemy>=2.0.23
alembic>=1.13.1

# Email
python-dotenv>=1.0.0
sendgrid>=6.11.0

# Google Sheets API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
gspread>=5.12.0
oauth2client>=4.1.3

# WhatsApp
twilio>=8.9.0

# Cache (optional)
redis>=5.0.0
rq>=1.15.1

# Monitoring & Error Tracking
sentry-sdk>=1.40.0

# CLI
rich>=13.7.0
click>=8.1.0

# Data Generation (for testing)
Faker>=22.0.0

# Utilities
requests>=2.31.0
httpx>=0.25.2
schedule>=1.2.1
apscheduler>=3.10.4
beautifulsoup4>=4.12.2
jsonschema>=4.20.0
PyJWT>=2.8.0
xxhash>=3.5.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-benchmark>=4.0.0
coverage>=7.4.0

# Code Quality
flake8>=7.0.0
bandit>=1.7.6
black>=24.1.1

# Security
bcrypt>=4.1.2
cryptography>=42.0.0
psycopg2-binary>=2.9.9

# Multimedia
openai-whisper>=20231117
pydub>=0.25.1
soundfile>=0.12.1

# Payments
mercadopago>=2.2.1
